fixed registration by nullabling companyId in user model. had the explicitly cast in a few other places to make up for it, no bigs.

+++ Controllers/HomeController.cs
@@ -55,7 +55,7 @@ namespace Unbugit.Controllers
             DashboardViewModel dashboard = new();
 
             BTUser currentUser = await _userManager.GetUserAsync(User);
-            int companyId = currentUser.CompanyId;
+            int companyId = (int)currentUser.CompanyId;
             if (User.IsInRole("Admin"))
             {
                 dashboard.Projects = await _projectService.GetAllProjectsByCompany(companyId);
